# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BCA Entertainment Works ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: getLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: assignedToId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lead list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /proposals:
    get:
      operationId: getProposals
      tags: [proposals]
      summary: List proposals
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: salespersonId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Proposal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proposal"
    post:
      operationId: createProposal
      tags: [proposals]
      summary: Create proposal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProposalBody"
      responses:
        "201":
          description: Created proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
...
  /proposals/{id}/won:
    post:
      operationId: markProposalWon
      tags: [proposals]
      summary: Mark proposal as Won and return work order stub
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal marked won with work order stub
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkWonResult"
  /proposals/{id}:
    get:
      operationId: getProposal
      tags: [proposals]
      summary: Get proposal by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    patch:
      operationId: updateProposal
      tags: [proposals]
      summary: Update proposal (including status transitions)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProposalBody"
      responses:
        "200":
          description: Updated proposal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proposal"
    delete:
      operationId: deleteProposal
      tags: [proposals]
      summary: Delete proposal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        company:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        gstNumber:
          type: string
          nullable: true
        billingAddress:
          type: string
          nullable: true
        deliveryAddress:
          type: string
          nullable: true
        source:
          type: string
          enum: [indiaMart, website, referral, direct, other]
        productInterest:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        lastFollowupNote:
          type: string
          nullable: true
        lastFollowupAt:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, contacted, proposalSent, negotiating, won, lost, onHold]
        assignedToId:
          type: integer
          nullable: true
        assignedToName:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerName
        - source
        - status
        - createdAt
        - updatedAt
    CreateLeadBody:
      type: object
      properties:
        customerName:
          type: string
        company:
          type: string
        phone:
          type: string
        email:
          type: string
        gstNumber:
          type: string
        billingAddress:
          type: string
        deliveryAddress:
          type: string
        source:
          type: string
          enum: [indiaMart, website, referral, direct, other]
        productInterest:
          type: string
        notes:
          type: string
        lastFollowupNote:
          type: string
        lastFollowupAt:
          type: string
        assignedToId:
          type: integer
      required:
        - customerName
    UpdateLeadBody:
      type: object
      properties:
        customerName:
          type: string
        company:
          type: string
        phone:
          type: string
        email:
          type: string
        gstNumber:
          type: string
        billingAddress:
          type: string
        deliveryAddress:
          type: string
        source:
          type: string
          enum: [indiaMart, website, referral, direct, other]
        productInterest:
          type: string
        notes:
          type: string
        lastFollowupNote:
          type: string
        lastFollowupAt:
          type: string
        status:
          type: string
          enum: [new, contacted, proposalSent, negotiating, won, lost, onHold]
        assignedToId:
          type: integer
...
    ProposalLineItem:
      type: object
      properties:
        description:
          type: string
        qty:
          type: number
        unitPrice:
          type: number
        gstRate:
          type: number
      required:
        - description
        - qty
        - unitPrice
    Proposal:
      type: object
      properties:
        id:
          type: integer
        proposalNumber:
          type: string
        leadId:
          type: integer
        customerName:
          type: string
          nullable: true
        company:
          type: string
          nullable: true
        salespersonId:
          type: integer
          nullable: true
        salespersonName:
          type: string
          nullable: true
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/ProposalLineItem"
        discountPercent:
          type: number
        gstRate:
          type: number
        subtotal:
          type: number
        discountAmount:
          type: number
        gstAmount:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [draft, sent, won, onHold, lost]
        validUntil:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        onHoldReminderDue:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - proposalNumber
        - leadId
        - lineItems
        - discountPercent
        - gstRate
        - subtotal
        - discountAmount
        - gstAmount
        - total
        - status
        - onHoldReminderDue
        - createdAt
        - updatedAt
    CreateProposalBody:
      type: object
      properties:
        leadId:
          type: integer
        salespersonId:
          type: integer
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/ProposalLineItem"
        discountPercent:
          type: number
        gstRate:
          type: number
        validUntil:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [draft, sent, won, onHold, lost]
      required:
        - leadId
    UpdateProposalBody:
      type: object
      properties:
        salespersonId:
          type: integer
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/ProposalLineItem"
        discountPercent:
          type: number
        gstRate:
          type: number
        validUntil:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [draft, sent, won, onHold, lost]
...
    WorkOrderStub:
      type: object
      properties:
        proposalId:
          type: integer
        proposalNumber:
          type: string
        leadId:
          type: integer
        customerName:
          type: string
          nullable: true
        total:
          type: number
        status:
          type: string
        message:
          type: string
      required:
        - proposalId
        - proposalNumber
        - leadId
        - total
        - status
        - message
    MarkWonResult:
      type: object
      properties:
        proposal:
          $ref: "#/components/schemas/Proposal"
        workOrder:
          $ref: "#/components/schemas/WorkOrderStub"
      required:
        - proposal
        - workOrder
...